+++ src/entities/tenant.js
@@ -30,6 +30,7 @@ class Tenant {
     paymentPurposeSuffix,
     applications,
     maxBookingAdvanceInMonths,
+    cancellationRefundTiers,
     defaultEventCreationMode,
     enablePublicStatusView,
     ownerUserIds,
@@ -67,6 +68,7 @@ class Tenant {
     this.paymentPurposeSuffix = paymentPurposeSuffix;
     this.applications = applications || [];
     this.maxBookingAdvanceInMonths = maxBookingAdvanceInMonths;
+    this.cancellationRefundTiers = cancellationRefundTiers || [];
     this.defaultEventCreationMode = defaultEventCreationMode || "";
     this.enablePublicStatusView = enablePublicStatusView;
     this.ownerUserIds = ownerUserIds || [];